apply max comments
