+++ policyengine_us/parameters/gov/states/ny/tax/income/credits/ctc/post_2024/older_child_age_min.yaml
@@ -1,0 +1,12 @@
+description: New York defines older children as being at or above this age for the Child Tax Credit amount, post 2024.
+values:
+  2000-01-01: 17  # Default to 17 (no eligibility)
+  2025-01-01: 4
+  2028-01-01: 17  # Reverts to old system post-2027
+metadata:
+  unit: year
+  period: year
+  label: New York CTC older child minimum age
+  reference:
+    - title: "New York Senate Bill S.3009-C, Part C, Section 2, Tax Law Section 606(c-1)(1-a)"
+      href: "https://assembly.state.ny.us/leg/?Text=Y&bn=S3009&default_fld=&leg_video=&term=2025"
