+++ policyengine_us/parameters/gov/states/ny/tax/income/credits/ctc/post_2024/phase_out/increment.yaml
@@ -1,0 +1,12 @@
+description: New York reduces the Child Tax Credit by a certain amount for each of this increment by which one's income exceeds the phase-out thresholds, post 2024.
+values:
+  2000-01-01: 1  # Default to 1 (no phase-out)
+  2025-01-01: 1_000
+  2028-01-01: 1  # Reverts to old system post-2027
+metadata:
+  unit: currency-USD
+  period: year
+  label: New York CTC phase-out increment
+  reference:
+    - title: "New York Senate Bill S.3009-C, Part C, Section 2, Tax Law Section 606(c-1)(1-a)(C)"
+      href: "https://assembly.state.ny.us/leg/?Text=Y&bn=S3009&default_fld=&leg_video=&term=2025"
