+++ policyengine_us/variables/gov/states/ny/tax/income/credits/ctc/ny_ctc_post_2024_phase_out.py
@@ -1,0 +1,38 @@
+from policyengine_us.model_api import *
+
+
+class ny_ctc_post_2024_phase_out(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = "NY CTC post-2024 phase-out amount"
+    documentation = "Amount by which NY CTC is reduced due to income phase-out under post-2024 rules"
+    unit = USD
+    definition_period = YEAR
+    reference = "https://www.nysenate.gov/legislation/laws/TAX/606"  # (c-1)
+    defined_for = StateCode.NY
+
+    def formula(tax_unit, period, parameters):
+        eligible = tax_unit("ny_ctc_post_2024_eligible", period)
+        if not eligible:
+            return 0
+
+        p = parameters(period).gov.states.ny.tax.income.credits.ctc
+        agi = tax_unit("adjusted_gross_income", period)
+        filing_status = tax_unit("filing_status", period)
+        base_credit = tax_unit("ny_ctc_post_2024_base", period)
+
+        # Only apply phase-out if there's a base credit and phase-out rate > 0
+        if (base_credit > 0) & (p.post_2024.phase_out.rate > 0):
+            phase_out_threshold = p.post_2024.phase_out.threshold[
+                filing_status
+            ]
+            excess_income = max_(agi - phase_out_threshold, 0)
+            # Round up to nearest increment for phase-out calculation
+            increment = p.post_2024.phase_out.increment
+            excess_increments = (excess_income + increment - 1) // increment
+            phase_out_amount = (
+                excess_increments * p.post_2024.phase_out.rate * increment
+            )
+            return phase_out_amount
+        else:
+            return 0
